+++ src/game/ghost.ts
@@ -1,26 +1,1 @@
-import { CameraFollowDef, setCameraFollowPosition } from "../camera.js";
-import { EM, EntityManager } from "../entity-manager.js";
-import { quat, vec3 } from "../gl-matrix.js";
-import { LinearVelocityDef } from "../physics/motion.js";
-import { PositionDef, RotationDef } from "../physics/transform.js";
-import { tempVec } from "../temp-pool.js";
-import { ControllableDef } from "./controllable.js";
 
-export const GhostDef = EM.defineComponent("ghost", () => ({}));
-
-export function createGhost(em: EntityManager) {
-  const g = em.newEntity();
-  em.ensureComponentOn(g, GhostDef);
-  em.ensureComponentOn(g, ControllableDef);
-  g.controllable.modes.canFall = false;
-  g.controllable.modes.canJump = false;
-  em.ensureComponentOn(g, CameraFollowDef, 1);
-  setCameraFollowPosition(g, "firstPerson");
-  em.ensureComponentOn(g, PositionDef);
-  em.ensureComponentOn(g, RotationDef);
-  // quat.rotateY(g.rotation, quat.IDENTITY, (-5 * Math.PI) / 8);
-  // quat.rotateX(g.cameraFollow.rotationOffset, quat.IDENTITY, -Math.PI / 8);
-  em.ensureComponentOn(g, LinearVelocityDef);
-
-  return g;
-}
