+++ src/game/sandbox.ts
@@ -1,0 +1,61 @@
+import {
+  CameraFollowDef,
+  setCameraFollowPosition,
+  CameraDef,
+} from "../camera.js";
+import { ColorDef } from "../color.js";
+import { EM, EntityManager } from "../entity-manager.js";
+import { vec3, quat } from "../gl-matrix.js";
+import { LinearVelocityDef } from "../physics/motion.js";
+import { PositionDef, RotationDef } from "../physics/transform.js";
+import { RenderableDef, RenderableConstructDef } from "../render/renderer.js";
+import { RendererDef } from "../render/render_init.js";
+import { AssetsDef } from "./assets.js";
+import { ControllableDef } from "./controllable.js";
+import { GlobalCursor3dDef } from "./cursor.js";
+
+export const GhostDef = EM.defineComponent("ghost", () => ({}));
+
+export function createGhost(em: EntityManager) {
+  const g = em.newEntity();
+  em.ensureComponentOn(g, GhostDef);
+  em.ensureComponentOn(g, ControllableDef);
+  g.controllable.modes.canFall = false;
+  g.controllable.modes.canJump = false;
+  em.ensureComponentOn(g, CameraFollowDef, 1);
+  setCameraFollowPosition(g, "firstPerson");
+  em.ensureComponentOn(g, PositionDef);
+  em.ensureComponentOn(g, RotationDef);
+  // quat.rotateY(g.rotation, quat.IDENTITY, (-5 * Math.PI) / 8);
+  // quat.rotateX(g.cameraFollow.rotationOffset, quat.IDENTITY, -Math.PI / 8);
+  em.ensureComponentOn(g, LinearVelocityDef);
+
+  return g;
+}
+export function initDbgGame(em: EntityManager, hosting: boolean) {
+  em.addSingletonComponent(CameraDef);
+
+  em.registerOneShotSystem(
+    null,
+    [AssetsDef, GlobalCursor3dDef, RendererDef],
+    (_, res) => {
+      const g = createGhost(em);
+      // em.ensureComponentOn(g, RenderableConstructDef, res.assets.cube.proto);
+      // createPlayer(em);
+
+      vec3.copy(g.position, [-16.6, 5, -5.1]);
+      quat.copy(g.rotation, [0, -0.77, 0, 0.636]);
+      quat.copy(g.cameraFollow.rotationOffset, [-0.225, 0, 0, 0.974]);
+
+      const c = res.globalCursor3d.cursor()!;
+      if (RenderableDef.isOn(c)) c.renderable.enabled = false;
+
+      vec3.copy(res.renderer.renderer.backgroundColor, [0.7, 0.8, 1.0]);
+
+      const p = em.newEntity();
+      em.ensureComponentOn(p, RenderableConstructDef, res.assets.plane.proto);
+      em.ensureComponentOn(p, ColorDef, [0.2, 0.3, 0.2]);
+      em.ensureComponentOn(p, PositionDef, [0, -5, 0]);
+    }
+  );
+}
